Add naïve counter implementation
This implementation should not be considered to be especially good. It
is a minimalistic busy spinning implementation.

The ESP8266 does not allow for long running tasks, especially IO tasks,
in timer callbacks.  For this reason, the programme is designed as a
busy loop with the only timer callback being the read_switches function.
This is NOT energetically optimal, but it works.

I propose that deep-sleeping and updating via pin-toggle interrupts
should be implemented in a future release.